Moved some stuff around.
