+++ include/tatami/subset/utils.hpp
@@ -2,7 +2,7 @@
 #define TATAMI_DELAYED_SUBSET_UTILS_HPP
 
 #include "../base/Matrix.hpp"
-#include "../base/new_extractor.hpp"
+#include "../utils/new_extractor.hpp"
 
 #include <vector>
 #include <algorithm>
